requirements.txt
streamlit==1.51.0
pandas==2.3.3
kagglehub==0.3.13
datasets==4.4.1
transformers==4.57.1
accelerate==1.11.0
sentencepiece==0.2.1
pillow==12.0.0
torch==2.9.1+cpu
torchvision==0.24.1+cpu
torchaudio==2.9.1+cpu
numpy==2.3.4
